Prevents :email being passed in to create_or_update!
When two contacts are merged in Hubspot, their emails are merged into a
comma-separated list. This list allows the contact to be found through
either email.

However, if the createOrUpdate endpoint matches on one of these
contacts and only one of the emails is passed in the properties hash,
Hubspot will attempt to update the email field. When using the email as
the identifier for a given Contact in createOrUpdate, the email should
not be included in the properties hash.