Address Travis test failures
Summary:
Due to a timezone issue, two tests in the `blog_spec` were passing
locally but failing consistently on Travis

The requests being issued to the `blog-posts` endpoint includes
a timestamp that limits the response to blog posts created in the
last two months. VCR records the timestamp in the URL and will
include this value when matching future requests against
recorded requests If VCR can't find a URL that matches the
request, a new real request is issued.

So why were the tests passing locally? The tests will pass locally
for anyone in the America/New_York timezone since the recorded
requests were generated using the America/New_York time zone.
Travis is failing on these tests because Travis uses the UTC
time zone.

This PR:
- Replaces `local` with `utc` to ensure consistency across time zones
- Updates the test assertions check the correct request is being made,
   rather than validating the API returned the expected values.
- Removes or updates tests marked as 'pending'
- Refactors to remove the use of `its` (which was deprecated with the
  release of RSpec 3) and `lets`