Update Contacts test coverage to use pre-recorded HTTP requests
Summary:
Creating an contact included a timestamp in the email address to ensure
creating a contact doesn't violate email constraints. Since the email
address is included in the HTTP request, VCR is unable to match the
request with a pre-recorded request. This results in VCR issuing a new
request each time this test is run.

This PR:
- Removes including a timestamp in the contact's email address so VCR
  can use a pre-recorded request.
- Removes the contacts integration test, which currently issues real
  network calls, in favor of the test coverage that already exists in
  the form of unit tests.
- Updates the tests for `.createOrUpdate` to be self-sufficient (creates
  necessary records, runs the expectations, deletes created records)
- introduces a contacts directory for VCR cassettes