Prevent VCR from making unexpected HTTP requests
Summary:
VCR defaults to using the record mode `:once`, which allows HTTP
interactions to be recorded when a cassette file doesn't exist but
raises when the cassette does exist and an unexpected request is made.

This behavior is currently being overridden in the `CassetteHelper` as
the record mode is being set to `:new_episodes`, which always allows new
HTTP interactions.

This PR:
- Sets the default record mode to `:none` to ensure new HTTP requests
  are intentional.
- Exposes the env variable VCR_RECORD to facilitate recording new HTTP
  interactions or updating existing cassettes.
- Removes references to the VCR record mode in individual tests
- Removes setting the record mode in the `CassetteHelper`