remove 'mock' in preparation of RSpec 3
Summary:
In preparation for upgrading to RSpec 3, this PR removes the use of the
test double library, [rr], in favor of RSpec's mocking framework. One of
the owners of the rr library has stated they do not have plans to
support RSpec 3. See the issue [Support for Rspec 3] for more details.

While RSpec 2 does expose a 'mock' method, RSpec 3 removes 'mock'
and 'stub' in favor of 'double' to create test doubles. Visit
the [RSpec changelog] for more details.

[rr]: https://github.com/rr/rr
[Support for RSpec 3]: https://github.com/rr/rr/issues/65
[RSpec changelog]:
https://github.com/rspec/rspec-mocks/blob/master/Changelog.md#300beta1--2013-11-07