Update CompanyProperty specs and reduce VCR yml file sizes
Summary:
This ticket experienced severe scope creep as I dove into the existing
tests for CompanyProperties. The original goal was to address to
particular test that wasn't executing the expectation because the
response was returning an empty list. While updating this test to
returns a non-empty response from the API, the API response was
returning over 1000 results, resulting in a VCR file that was over 1MB.
The response is so large because we're using the `demo` account which is
used by everyone testing out the HubSpot API.

To resolve this concern, this PR pulls in the `dotenv` library so the
`hapikey` can be set to a value other than `demo`. In my case, I
followed HubSpot's documentation for creating a [Developer Account] and
used the Portal hapikey. This resulted in reducing the VCR recording
that was 1.3MB down to 136KB. With time, I think we'll want to replace
using the `hapikey` with an OAuth Access Token, as recommended by the
HubSpot in the  Authentication API docs.

[Developer Account]:
https://developers.hubspot.com/docs/methods/auth/oauth-overview

A few other items this PR addresses:
- Removes trailing slashes from the CompanyProperty paths
- Updates VCR to filter out the hapikey so it's not stored in the
  recording.